+++ test/textures.rb
@@ -1,4 +1,20 @@
 class TestTexture < MTest::Unit::TestCase
+  def test_load_texture_from_image
+    skip_unless_display_present
+
+    set_window_title(__method__.to_s)
+
+    image = Image.load('test/assets/test.png')
+    texture = load_texture_from_image(image)
+    clear_and_draw do
+      draw_texture(texture, 3, 3, WHITE)
+    end
+
+    assert_equal fixture_draw_texture, get_screen_data.data
+    texture.unload
+    image.unload
+  end
+
   def test_draw_texture
     skip_unless_display_present